Refactor: rename experiment.py to experiment_setup.py
