+++ examples/synthetic_graph_evolution/graph_search.py
@@ -2,7 +2,7 @@ from datetime import timedelta
 from functools import partial
 from typing import Type, Optional
 
-from examples.synthetic_graph_evolution.experiment import run_experiments
+from examples.synthetic_graph_evolution.experiment_setup import run_experiments
 from examples.synthetic_graph_evolution.generators import generate_labeled_graph
 from golem.core.adapter.nx_adapter import BaseNetworkxAdapter
 from golem.core.dag.verification_rules import has_no_self_cycled_nodes
